@@ -1,54 +1,59 @@
 {
-  "name": "sample-app",
-  "version": "0.0.0",
-  "license": "MIT",
+  "name": "angular-io-example",
+  "version": "1.0.0",
+  "private": true,
+  "description": "Example project from an angular.io guide.",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build",
     "test": "ng test",
-    "lint": "ng lint",
+    "lint": "tslint ./src/**/*.ts -t verbose",
     "e2e": "ng e2e"
   },
-  "private": true,
+  "keywords": [],
+  "author": "",
+  "license": "MIT",
   "dependencies": {
-    "@angular/animations": "^4.2.4",
-    "@angular/common": "^4.2.4",
-    "@angular/compiler": "^4.2.4",
-    "@angular/core": "^4.2.4",
-    "@angular/forms": "^4.2.4",
-    "@angular/http": "^4.2.4",
-    "@angular/platform-browser": "^4.2.4",
-    "@angular/platform-browser-dynamic": "^4.2.4",
-    "@angular/router": "^4.2.4",
-    "@ngrx/core": "^1.2.0",
+    "@angular/animations": "^5.0.1",
+    "@angular/common": "^5.0.1",
+    "@angular/compiler": "^5.0.1",
+    "@angular/compiler-cli": "^5.0.1",
+    "@angular/core": "^5.0.1",
+    "@angular/forms": "^5.0.1",
+    "@angular/http": "^5.0.1",
+    "@angular/platform-browser": "^5.0.1",
+    "@angular/platform-browser-dynamic": "^5.0.1",
+    "@angular/platform-server": "^5.0.1",
+    "@angular/router": "^5.0.1",
+    "@angular/upgrade": "~5.0.0",
     "@ngrx/effects": "^2.0.5",
-    "@ngrx/store": "^2.2.3",
-    "angular-in-memory-web-api": "0.0.10",
+    "@ngrx/store": "^4.1.1",
+    "angular-in-memory-web-api": "~0.5.0",
     "core-js": "^2.4.1",
-    "lodash": "^4.17.4",
-    "rxjs": "^5.4.2",
-    "zone.js": "^0.8.14"
+    "rxjs": "^5.5.0",
+    "typescript": "^2.6.1",
+    "web-animations-js": "^2.3.1",
+    "zone.js": "^0.8.4"
   },
   "devDependencies": {
-    "@angular/cli": "1.4.9",
-    "@angular/compiler-cli": "^4.2.4",
-    "@angular/language-service": "^4.2.4",
+    "@angular/cli": "1.5.0",
     "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.2.0",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "@types/jasminewd2": "^2.0.3",
+    "@types/node": "^6.0.45",
+    "jasmine": "~2.4.1",
+    "jasmine-core": "~2.4.1",
+    "karma": "^1.3.0",
+    "karma-chrome-launcher": "^2.0.0",
+    "karma-cli": "^1.0.1",
+    "karma-jasmine": "^1.0.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.3.3"
-  }
+    "karma-phantomjs-launcher": "^1.0.2",
+    "lodash": "^4.16.2",
+    "phantomjs-prebuilt": "^2.1.7",
+    "protractor": "~5.1.0",
+    "ts-node": "^3.3.0",
+    "tslint": "^3.15.1",
+    "typescript": "2.4.2"
+  },
+  "repository": {}
 }
