@@ -1,10 +1,11 @@
 import { Component, OnInit, Input } from '@angular/core';
-import { Hero } from './hero';
-import { HeroService} from './hero.service';
+import { Hero } from '../hero';
+import { HeroService} from '../hero.service';
 import { Router } from '@angular/router';
-// import { Observable } from 'rxjs/Observable';
-// import { Store } from '@ngrx/store';
-// import { AppState } from './reducers/index';
+import { AppState } from '../reducers/index.reducer';
+import { Observable } from 'rxjs/Observable';
+import { Store } from '@ngrx/store';
+
 
 
 @Component({
@@ -13,13 +14,14 @@ import { Router } from '@angular/router';
   styleUrls: ['./heroes.component.css'],
 })
 
-export class HeroesComponent implements OnInit {
+export class HeroesComponent implements OnInit{
 
   heroes: Hero[];
   selectedHero: Hero;
   oneWayBinding = 'One Way Binding';
 
-  constructor(private heroService: HeroService, private router: Router) {}
+  constructor(private heroService: HeroService, private router: Router) {
+  }
 
   getHeroes(): void {
     // this.heroes = this.store.select('heroes');
