@@ -1,5 +1,5 @@
 import { TestBed, async } from '@angular/core/testing';
-import { HeroesComponent } from './heroes.component';
+import { HeroesComponent } from './heroes/heroes.component';
 describe('AppComponent', () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
