@@ -1,5 +1,6 @@
 import { Injectable } from '@angular/core';
-import { Effect, Actions } from '@ngrx/effects';
+import { Effect, Actions, toPayload } from '@ngrx/effects';
+import { AppState } from '../reducers/index.reducer';
 import { HeroActions } from '../actions/hero-actions.component';
 import 'rxjs/add/operator/map';
 import { HeroService } from '../hero.service';
