@@ -1,79 +1,121 @@
 import { Injectable } from '@angular/core';
-import { HttpClient, HttpHeaders} from '@angular/common/http';
-import { MessageService } from './message.service';
-import { Observable } from 'rxjs/Observable';
-import { filter } from 'rxjs/operator/filter';
-import { catchError, map, tap } from 'rxjs/operators';
-import { Hero } from './hero';
-import { HEROES } from './mock-heroes';
-import 'rxjs/add/operator/toPromise';
-import { HERO_DATA } from './hero-data';
-import { of } from 'rxjs/observable/of';
-import * as _ from 'lodash';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 
+import { Observable } from 'rxjs/Observable';
+import { of } from 'rxjs/observable/of';
+import { catchError, map, tap } from 'rxjs/operators';
+
+import { Hero } from './hero';
+import { MessageService } from './message.service';
+
+const httpOptions = {
+  headers: new HttpHeaders({ 'Content-Type': 'application/json' })
+};
 
 @Injectable()
-
 export class HeroService {
-    private heroesUrl = 'api/heroes';  // URL to web api
-    private heroData: Hero[] = [];
 
-    constructor(private http: HttpClient, private  messageService: MessageService) {
-        this.heroData = [...HERO_DATA];
+  private heroesUrl = 'api/heroes';  // URL to web api
+
+  constructor(
+    private http: HttpClient,
+    private messageService: MessageService) { }
+
+  /** GET heroes from the server */
+  getHeroes (): Observable<Hero[]> {
+    return this.http.get<Hero[]>(this.heroesUrl)
+      .pipe(
+        tap(heroes => this.log(`fetched heroes`)),
+        catchError(this.handleError('getHeroes', []))
+      );
+  }
+
+  /** GET hero by id. Return `undefined` when id not found */
+  getHeroNo404<Data>(id: number): Observable<Hero> {
+    const url = `${this.heroesUrl}/?id=${id}`;
+    return this.http.get<Hero[]>(url)
+      .pipe(
+        map(heroes => heroes[0]), // returns a {0|1} element array
+        tap(h => {
+          const outcome = h ? `fetched` : `did not find`;
+          this.log(`${outcome} hero id=${id}`);
+        }),
+        catchError(this.handleError<Hero>(`getHero id=${id}`))
+      );
+  }
+
+  /** GET hero by id. Will 404 if id not found */
+  getHero(id: number): Observable<Hero> {
+    const url = `${this.heroesUrl}/${id}`;
+    return this.http.get<Hero>(url).pipe(
+      tap(_ => this.log(`fetched hero id=${id}`)),
+      catchError(this.handleError<Hero>(`getHero id=${id}`))
+    );
+  }
+
+  /* GET heroes whose name contains search term */
+  searchHeroes(term: string): Observable<Hero[]> {
+    if (!term.trim()) {
+      // if not search term, return empty hero array.
+      return of([]);
     }
+    return this.http.get<Hero[]>(`api/heroes/?name=${term}`).pipe(
+      tap(_ => this.log(`found heroes matching "${term}"`)),
+      catchError(this.handleError<Hero[]>('searchHeroes', []))
+    );
+  }
 
-    // getHeroes(): Promise<Hero[]> {
-    //     return Promise.resolve(HEROES);
-    // }
+  //////// Save methods //////////
 
-    getHeroes(): Observable<Hero[]> {
-        return of(HERO_DATA);
-        // return this.http.get<Hero[]>(this.heroesUrl)
-        //     .pipe(
-        //         tap(heroes => this.log(`fetched heroes`)),
-        //         catchError(this.handleError('getHeroes', []))
-        //     );
-    }
+  /** POST: add a new hero to the server */
+  addHero (hero: Hero): Observable<Hero> {
+    return this.http.post<Hero>(this.heroesUrl, hero, httpOptions).pipe(
+      tap((hero: Hero) => this.log(`added hero w/ id=${hero.id}`)),
+      catchError(this.handleError<Hero>('addHero'))
+    );
+  }
 
-    getHero(id: number): Observable<Hero> {
-        return of(_.find(HERO_DATA, 'id', id));
-        // const url = `${this.heroesUrl}/?id=${id}`;
-        // return this.http.get<Hero>(url).pipe(
-        //     tap(_ => this.log(`fetched hero id=${id}`)),
-        //     catchError(this.handleError<Hero>(`getHero id=${id}`))
-        // );
-    }
+  /** DELETE: delete the hero from the server */
+  deleteHero (hero: Hero | number): Observable<Hero> {
+    const id = typeof hero === 'number' ? hero : hero.id;
+    const url = `${this.heroesUrl}/${id}`;
 
-    // getHero(id: number): Promise<Hero> {
-    //     return this.getHeroes().then(heroes => heroes.find(hero => hero.id === id));
-    // }
+    return this.http.delete<Hero>(url, httpOptions).pipe(
+      tap(_ => this.log(`deleted hero id=${id}`)),
+      catchError(this.handleError<Hero>('deleteHero'))
+    );
+  }
 
-    // private handleError(error: any): Promise<any> {
-    //     console.error('An error occurred', error); // for demo purposes only
-    //     return Promise.reject(error.message || error);
-    // }
+  /** PUT: update the hero on the server */
+  updateHero (hero: Hero): Observable<any> {
+    return this.http.put(this.heroesUrl, hero, httpOptions).pipe(
+      tap(_ => this.log(`updated hero id=${hero.id}`)),
+      catchError(this.handleError<any>('updateHero'))
+    );
+  }
 
-    private log(message: string) {
-        this.messageService.add('HeroService: ' + message);
-    }
+  /**
+   * Handle Http operation that failed.
+   * Let the app continue.
+   * @param operation - name of the operation that failed
+   * @param result - optional value to return as the observable result
+   */
+  private handleError<T> (operation = 'operation', result?: T) {
+    return (error: any): Observable<T> => {
 
-    private handleError<T> (operation = 'operation', result?: T) {
-        return (error: any): Observable<T> => {
-            console.error(error);
+      // TODO: send the error to remote logging infrastructure
+      console.error(error); // log to console instead
 
-            this.log(`${operation} failed: ${error.message}`);
-            return of(result as T);
-        };
-    }
+      // TODO: better job of transforming error for user consumption
+      this.log(`${operation} failed: ${error.message}`);
 
+      // Let the app keep running by returning an empty result.
+      return of(result as T);
+    };
+  }
 
-    /* for HTTP request and using observables */
-    // getHero(id: number): Promise<Hero> {
-    //     const url = `${this.heroesUrl}/${id}`;
-    //     return this.http.get(url)
-    //         .toPromise()
-    //         .then(response => response.json().data as Hero)
-    //         .catch(this.handleError);
-    // }
-
+  /** Log a HeroService message with the MessageService */
+  private log(message: string) {
+    this.messageService.add('HeroService: ' + message);
+  }
 }
