@@ -1,6 +1,7 @@
-import { InMemoryDbService } from 'angular2-in-memory-web-api';
+import { InMemoryDbService } from 'angular-in-memory-web-api';
 
-export class InMemoryDataService implements InMemoryDbService {
+export class InMemoryDataService implements InMemoryDbService
+{
     createDb() {
         const heroes = [
             {id: 11, name: 'Black Widow', strength : 7},
